First Time Commit.
This is my repo of haxe projects.
My actual haxe libraries are in a mercurial bitbucket repo since you can
make bitbucket private. It is nice to have the project code in the same
repo... But he projects take up a lot of memory. Where the actual
library code is pretty small.